+++ account-kit/signer/src/client/types.ts
@@ -95,6 +95,19 @@ export type OtpResponse =
       multiFactors: MfaFactor[];
     };
 
+export type JwtParams = {
+  jwt: string;
+  targetPublicKey: string;
+  authProvider?: string;
+  expirationSeconds?: number;
+};
+
+export type JwtResponse = {
+  isSignUp: boolean;
+  orgId: string;
+  credentialBundle: string;
+};
+
 export type SignupResponse = {
   orgId: string;
   userId?: string;
@@ -264,6 +277,11 @@ export type SignerEndpoints = [
     };
   },
   {
+    Route: "/v1/auth-jwt";
+    Body: JwtParams;
+    Response: JwtResponse;
+  },
+  {
     Route: "/v1/signer-config";
     Body: {};
     Response: SignerConfig;
@@ -344,10 +362,43 @@ export type SignerEndpoints = [
       };
     };
   },
+  {
+    Route: "/v1/multi-owner-prepare-delete";
+    Body: {
+      organizationId: string;
+      members: {
+        evmSignerAddress: Address;
+      }[];
+    };
+    Response: {
+      result: {
+        updateRootQuorumRequest: TurnkeyApiTypes["v1UpdateRootQuorumRequest"];
+        deleteMembersRequest: TurnkeyApiTypes["v1DeleteUsersRequest"];
+      };
+    };
+  },
+  {
+    Route: "/v1/multi-owner-delete";
+    Body: {
+      stampedRequest: TSignedRequest;
+    };
+    Response: {
+      result: {
+        deletedUserIds: string[];
+      };
+    };
+  },
 ];
 
 export type AuthenticatingEventMetadata = {
-  type: "email" | "passkey" | "oauth" | "otp" | "otpVerify" | "sms";
+  type:
+    | "email"
+    | "passkey"
+    | "oauth"
+    | "otp"
+    | "otpVerify"
+    | "custom-jwt"
+    | "sms";
 };
 
 export type AlchemySignerClientEvents = {
@@ -358,6 +409,7 @@ export type AlchemySignerClientEvents = {
   connectedPasskey(user: User): void;
   connectedOauth(user: User, bundle: string): void;
   connectedOtp(user: User, bundle: string): void;
+  connectedJwt(user: User, bundle: string): void;
   disconnected(): void;
 };